pyproject.toml
[project]
name = "self-rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.3.0",
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "qdrant-client>=1.12.0",
    "alembic>=1.13.2",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.7",
    "aiosqlite>=0.19.0",
    "rank-bm25>=0.2.2",
    "llama-index-core>=0.11.0",
    "llama-index-vector-stores-qdrant>=0.2.0",
    "llama-index-retrievers-bm25>=0.1.4",
    "llama-index-llms-openai>=0.1.26",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app", "app.*", "sdk", "sdk.*"]
exclude = ["alembic", "alembic.*", "tests", "tests.*"]
